add 2 exercicios

+++ prog1/ex-for/ex012.py
@@ -1,7 +1,7 @@
 # Escreva um programa que leia 5 pares de 2 valores, o primeiro representando a matrícula de um aluno e o segundo representando a sua altura em centímetros. Encontrar o aluno mais alto e o mais baixo e escrever suas matrículas, suas alturas e uma mensagem dizendo se é o mais alto ou o mais baixo.
 
 matricula = int(input('Digite a matrícula do aluno: '))
-altura = int(input('Digite a altura do respectivo aluno: '))
+altura = float(input('Digite a altura do respectivo aluno: '))
 maisAlto = altura
 matriculaMaisAlto = matricula
 maisBaixo = altura
@@ -10,7 +10,7 @@ matriculaMaisBaixo = matricula
 
 for i in range (4):
     matricula = int(input('Digite a matrícula do aluno: '))
-    altura = int(input('Digite a altura do respectivo aluno: '))
+    altura = float(input('Digite a altura do respectivo aluno: '))
 
     if altura > maisAlto:
         maisAlto = altura
